Separate server from app main point
